api/pyproject.toml
[tool.poetry]
name = "graphrag-agent"
version = "0.1.0"
description = ""
authors = ["Benjamin Chan"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.2.2"}
pydantic = "<2"
langchain-community = "^0.2.5"
langchain-core = "^0.2.7"
neo4j = "^5.20.0"
langchain-openai = "^0.1.8"
langchain-text-splitters = "^0.2.1"
tiktoken = "^0.7.0"
langchain-experimental = "^0.0.61"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
